+++ KNUTICE/Presentation/WebView/WebViewController.swift
@@ -14,6 +14,7 @@ final class WebViewController: UIViewController {
     let progressView: UIProgressView = {
         let progressView = UIProgressView(progressViewStyle: .bar)
         progressView.progressTintColor = .accent2
+        progressView.translatesAutoresizingMaskIntoConstraints = false
         
         return progressView
     }()
@@ -24,6 +25,7 @@ final class WebViewController: UIViewController {
         webView.allowsBackForwardNavigationGestures = false
         webView.configuration.defaultWebpagePreferences.allowsContentJavaScript = true
         webView.isHidden = true
+        webView.translatesAutoresizingMaskIntoConstraints = false
         
         return webView
     }()
@@ -41,6 +43,7 @@ final class WebViewController: UIViewController {
         button.layer.shadowRadius = 7
         button.layer.shadowOffset = CGSize(width: 0, height: 0)
         button.addTarget(self, action: #selector(openReminderForm(_:)), for: .touchUpInside)
+        button.translatesAutoresizingMaskIntoConstraints = false
         
         return button
     }()
@@ -121,14 +124,12 @@ extension WebViewController {
     func setupLayout() {
         //progressView
         view.addSubview(progressView)
-        progressView.translatesAutoresizingMaskIntoConstraints = false
         progressView.snp.makeConstraints { make in
             make.left.top.right.equalTo(self.view.safeAreaLayoutGuide)
         }
         
         //webView
         view.addSubview(webView)
-        webView.translatesAutoresizingMaskIntoConstraints = false
         webView.snp.makeConstraints { make in
             make.top.equalTo(progressView.snp.bottom)
             make.leading.trailing.bottom.equalToSuperview()
@@ -136,12 +137,46 @@ extension WebViewController {
         
         //reminderSheetBtn
         view.addSubview(reminderSheetBtn)
-        reminderSheetBtn.translatesAutoresizingMaskIntoConstraints = false
         reminderSheetBtn.snp.makeConstraints { make in
             make.bottom.equalToSuperview().offset(-50)
             make.trailing.equalToSuperview().offset(-20)
             make.width.height.equalTo(50)
         }
+    }
+}
+
+extension WebViewController {
+    @objc func openSharePanel() {
+        let shareText = self.notice.contentUrl
+        let shareObject = [shareText]
+        let activityViewController = UIActivityViewController(activityItems : shareObject, applicationActivities: nil)
+        activityViewController.popoverPresentationController?.sourceView = self.view
+        activityViewController.completionWithItemsHandler = { (activityType: UIActivity.ActivityType?, completed: Bool, arrayReturnedItems: [Any]?, error: Error?) in
+            if completed {
+                self.showCompleAlert()
+            } else {
+                
+            }
+        }
+        self.present(activityViewController, animated: true, completion: nil)
+    }
+    
+    private func showCompleAlert() {
+        let alert = UIAlertController(title: "알림", message: "공유를 완료했어요.", preferredStyle: .actionSheet)
+        let okButton = UIAlertAction(title: "확인", style: .default, handler: nil)
+        alert.addAction(okButton)
+        present(alert, animated: true, completion: nil)
+        
+    }
+    
+    @objc func openReminderForm(_ sender: UIButton) {
+        let bookmarkForm = BookmarkForm(viewModel: BookmarkFormViewModel(), notice: self.notice) {
+            self.dismiss(animated: true)
+        }
+        let viewController = UIHostingController(rootView: bookmarkForm)
+        let navigationController = UINavigationController(rootViewController: viewController)
+        navigationController.modalPresentationStyle = .popover
+        present(navigationController, animated: true, completion: nil)
     }
 }
